# backend/static/openapi.yaml
openapi: "3.0.3"
info:
  title: "IOT Smart Home for Elderly — Alfred API"
  description: |
    REST API cho hệ thống nhà thông minh hỗ trợ người cao tuổi.
    
    ## Xác thực
    Tất cả endpoint (trừ `/api/auth/login`, `/api/auth/register`) đều yêu cầu signed bearer token.
    Token được ký bằng `itsdangerous` và được gửi qua **HttpOnly Cookie** (`batman_os_auth`) hoặc header `Authorization: Bearer <token>`.
    
    ## Rate Limiting
    - Login: 5 req/phút, 20 req/giờ
    - AI Chat: 30 req/phút, 300 req/giờ
    - Devices: 60 req/phút, 2000 req/giờ
    - Mặc định: 100 req/giờ, 500 req/ngày
  version: "2.0.0"
  contact:
    name: "Tu Dinh Huy — ITITIU21213"
    email: "ITITIU21213@student.hcmiu.edu.vn"
  license:
    name: "Academic (Thesis Project)"
servers:
  - url: "http://127.0.0.1:5000"
    description: "Local development"
  - url: "https://your-domain.com"
    description: "Production"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: batman_os_auth
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    Error:
      type: object
      properties:
        status:
          type: string
          example: "error"
        message:
          type: string
          example: "Authentication required"
...
    Device:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Đèn phòng khách"
        code:
          type: string
          example: "den"
        icon:
          type: string
          example: "💡"
        category:
          type: string
          enum: [light, fan, sensor, ac, other]
          example: "light"
        room_id:
          type: integer
          nullable: true
          example: 2
        status:
          type: object
          properties:
            state:
              type: string
              enum: [on, off, unknown]
              example: "off"
            updated_at:
              type: string
              format: date-time
        map_x:
          type: number
          nullable: true
        map_y:
          type: number
          nullable: true
...
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
...
  /api/devices:
    get:
      tags: [devices]
      summary: Lấy danh sách thiết bị
      parameters:
        - in: query
          name: room_id
          schema:
            type: integer
          description: Lọc theo phòng
      responses:
        "200":
          description: Danh sách thiết bị
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  status:
                    type: string
                    example: "online"
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Device"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      tags: [devices]
      summary: Tạo thiết bị mới
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, code]
              properties:
                name:
                  type: string
                  example: "Đèn phòng ngủ"
                code:
                  type: string
                  example: "den_pn"
                  description: Mã duy nhất của thiết bị, phải khớp với firmware ESP32
                icon:
                  type: string
                  example: "💡"
                category:
                  type: string
                  enum: [light, fan, sensor, ac, other]
                  example: "light"
                room_id:
                  type: integer
                  nullable: true
                control_types:
                  type: array
                  items:
                    type: string
                  example: ["on", "off"]
      responses:
        "201":
          description: Thiết bị đã được tạo
        "400":
          description: Dữ liệu không hợp lệ
  /api/devices/control:
    post:
      tags: [devices]
      summary: Điều khiển thiết bị (bật/tắt)
      description: |
        Gửi lệnh đến thiết bị qua MQTT. Backend publish message đến `home/control/{code}`.
        ESP32 nhận và thực hiện lệnh, sau đó phản hồi qua `home/status/{code}`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [device_name, action]
              properties:
                device_name:
                  type: string
                  example: "Đèn phòng khách"
                action:
                  type: string
                  enum: [on, off, toggle]
                  example: "on"
      responses:
        "200":
          description: Lệnh đã được gửi
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
  /api/devices/{id}:
    get:
      tags: [devices]
      summary: Lấy chi tiết một thiết bị
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Thông tin thiết bị
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "404":
          description: Không tìm thấy thiết bị

    put:
      tags: [devices]
      summary: Cập nhật thông tin thiết bị
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                room_id:
                  type: integer
                  nullable: true
                icon:
                  type: string
                category:
                  type: string
      responses:
        "200":
          description: Cập nhật thành công

    delete:
      tags: [devices]
      summary: Xóa thiết bị (soft delete)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Xóa thành công
        "404":
          description: Không tìm thấy
  /api/devices/{id}/history:
    get:
      tags: [devices]
      summary: Lịch sử điều khiển thiết bị
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Danh sách lịch sử điều khiển
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        action:
                          type: string
                        timestamp:
                          type: string
                          format: date-time
                        source:
                          type: string
                          enum: [app, mqtt, schedule, ai]

  # ================================================================
  # ROOMS
  # ================================================================
...
components:
  responses:
    Unauthorized:
      description: Token không hợp lệ hoặc hết hạn
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"